+++ internal/line-highlight-plugin.js
@@ -18,9 +18,9 @@ export function LineHighlightPlugin (options) {
     let index = 1
     for (const token of env.tokens) {
       if (typeof token === "string") continue
-      if (!token.type.includes("light-line")) continue
 
       if (highlightLinesRange.includes(index)) {
+        // line-highlight is what prism themes expect.
         token.type += " line-highlight"
       }
 
@@ -32,7 +32,55 @@ export function LineHighlightPlugin (options) {
         token.type += " deleted"
       }
 
-      index++;
+      // Tokens can only ever be "light-line" or "light-gutter-cell" so we just increment on "light-line"
+      if (token.type.includes("light-line")) {
+        index++;
+      }
     }
   }
+}
+
+/**
+  * @typedef {object} WrapEnv
+  * @property {string} type
+  * @property {string} content
+  * @property {Array<string>} classes
+  * @property {Record<string, string>} attributes
+  * @property {string} language
+  */
+
+
+/**
+ * @example
+ *   prism.hooks.add("wrap", LineHighlightWrapPlugin)
+ */
+export function LineHighlightWrapPlugin () {
+  /**
+    * @param {WrapEnv} env
+    */
+  return function (env) {
+    const partTypes = ["inserted", "line-highlight", "deleted"]
+
+    if (env.type.includes('light-line')) {
+	    env.attributes['part'] = "line"
+
+      partTypes.forEach((type) => {
+        if (!env.type.includes(type)) return
+        if (type === "line-highlight") type = "highlight"
+
+        env.attributes["part"] += " line-" + type
+      })
+    }
+
+    if (env.type.includes("light-gutter-cell")) {
+	    env.attributes['part'] = "gutter-cell"
+
+      partTypes.forEach((type) => {
+        if (!env.type.includes("gutter-cell-" + type)) return
+        if (type === "line-highlight") type = "highlight"
+
+        env.attributes["part"] += " gutter-cell-" + type
+      })
+	  }
+	}
 }